// frontend/src/app/api/generated/model/solarium-account.ts
/**
 * Excalibur Gym API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SolariumTransaction } from './solarium-transaction';


export interface SolariumAccount { 
    guestId: string;
    guestName: string;
    remainingMinutes: number;
    transactions: Array<SolariumTransaction>;
}