// frontend/src/app/api/generated/model/solarium-payment-method.ts
/**
 * Excalibur Gym API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum SolariumPaymentMethod {

    Cash = 'CASH',

    BankCard = 'BANK_CARD',

    PrepaidBalance = 'PREPAID_BALANCE'
}