// frontend/src/app/api/generated/model/solarium-product.ts
/**
 * Excalibur Gym API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SolariumProduct { 
    id: string;
    code: string;
    name: string;
    minutesPerUnit: number;
    defaultPrice: number;
    active: boolean;
    updatedAt: string;
}