// frontend/src/app/api/generated/model/solarium-transaction.ts
/**
 * Excalibur Gym API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SolariumPaymentMethod } from './solarium-payment-method';


export interface SolariumTransaction { 
    id: string;
    type: SolariumTransactionTypeEnum;
    minutesDelta: number;
    balanceAfter: number;
    productName?: string | null;
    quantity?: number | null;
    pricePaid?: number | null;
    paymentMethod?: SolariumPaymentMethod;
    occurredAt: string;
    employeeName: string;
}
export enum SolariumTransactionTypeEnum {
    Purchase = 'PURCHASE',
    Usage = 'USAGE'
};